+++ airflow/providers/apache/druid/operators/druid_check.py
@@ -17,19 +17,19 @@
 # under the License.
 import warnings
 
-from airflow.operators.sql import SQLCheckOperator
+from airflow.providers.common.sql.operators.sql import SQLCheckOperator
 
 
 class DruidCheckOperator(SQLCheckOperator):
     """
     This class is deprecated.
-    Please use `airflow.operators.sql.SQLCheckOperator`.
+    Please use `airflow.providers.common.sql.operators.sql.SQLCheckOperator`.
     """
 
     def __init__(self, druid_broker_conn_id: str = 'druid_broker_default', **kwargs):
         warnings.warn(
             """This class is deprecated.
-            Please use `airflow.operators.sql.SQLCheckOperator`.""",
+            Please use `airflow.providers.common.sql.operators.sql.SQLCheckOperator`.""",
             DeprecationWarning,
             stacklevel=2,
         )
