+++ airflow/providers/snowflake/provider.yaml
@@ -44,7 +44,7 @@ versions:
 
 dependencies:
   - apache-airflow>=2.2.0
-  - apache-airflow-providers-common-sql
+  - apache-airflow-providers-common-sql>=1.1.0
   - snowflake-connector-python>=2.4.1
   - snowflake-sqlalchemy>=1.1.0
