+++ airflow/sensors/sql_sensor.py
@@ -15,7 +15,7 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
-"""This module is deprecated. Please use :mod:`airflow.sensors.sql`."""
+"""This module is deprecated. Please use :mod:`airflow.providers.common.sql.sensors.sql`."""
 
 import warnings
