+++ docs/apache-airflow/operators-and-hooks-ref.rst
@@ -77,9 +77,6 @@ For details see: :doc:`apache-airflow-providers:operators-and-hooks-ref/index`.
    * - :mod:`airflow.operators.python`
      - :doc:`How to use <howto/operator/python>`
 
-   * - :mod:`airflow.operators.sql`
-     -
-
    * - :mod:`airflow.operators.subdag`
      -
 
@@ -111,9 +108,6 @@ For details see: :doc:`apache-airflow-providers:operators-and-hooks-ref/index`.
 
    * - :mod:`airflow.sensors.smart_sensor`
      - :doc:`concepts/smart-sensors`
-
-   * - :mod:`airflow.sensors.sql`
-     -
 
    * - :mod:`airflow.sensors.time_delta`
      -
